Move ProgramContext to executor
ProgramContext is only used at runtime, so it belongs in smite-scenarios
with the executor rather than in smite-ir with the IR.
