+++ smite-scenarios/src/executor.rs
@@ -9,7 +9,21 @@ use smite::bolt::{
 };
 use smite::noise::{ConnectionError, NoiseConnection};
 use smite_ir::operation::AcceptChannelField;
-use smite_ir::{Operation, Program, ProgramContext, Variable, VariableType};
+use smite_ir::{Operation, Program, Variable, VariableType};
+
+/// State captured during snapshot setup, available to IR programs at execution
+/// time via `LoadContext*` operations.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ProgramContext {
+    /// Target node's compressed public key.
+    pub target_pubkey: [u8; 33],
+    /// Chain hash (genesis block hash).
+    pub chain_hash: [u8; 32],
+    /// Current block height at snapshot time.
+    pub block_height: u32,
+    /// Target's advertised feature bits from init message.
+    pub target_features: Vec<u8>,
+}
 
 /// Abstraction over a Noise-encrypted connection, allowing mock implementations
 /// in tests.
